+++ pmd-apex/src/main/java/net/sourceforge/pmd/lang/apex/ast/ASTUserClass.java
@@ -4,8 +4,12 @@
 
 package net.sourceforge.pmd.lang.apex.ast;
 
+import java.util.stream.Collectors;
+
 import net.sourceforge.pmd.Rule;
 
+import apex.jorje.data.Identifier;
+import apex.jorje.data.ast.TypeRef;
 import apex.jorje.semantic.ast.compilation.UserClass;
 
 public class ASTUserClass extends ApexRootNode<UserClass> implements ASTUserClassOrInterface<UserClass>,
@@ -66,5 +70,12 @@ public class ASTUserClass extends ApexRootNode<UserClass> implements ASTUserClas
 
     public ASTModifierNode getModifiers() {
         return getFirstChildOfType(ASTModifierNode.class);
+    }
+
+
+    public String getSuperTypeName() {
+        return node.getDefiningType().getCodeUnitDetails().getSuperTypeRef().map(TypeRef::getNames)
+            .map(it -> it.stream().map(Identifier::getValue).collect(Collectors.joining(".")))
+            .orElse("");
     }
 }