+++ pmd-apex/src/test/java/net/sourceforge/pmd/lang/apex/ast/ASTUserClassTest.java
@@ -28,4 +28,20 @@ public class ASTUserClassTest {
         Assert.assertNotNull(innerNode);
         Assert.assertEquals("Bar", innerNode.getImage());
     }
+
+    @Test
+    public void testSuperClassType() {
+        ApexNode<?> node = parse("public class AccountTriggerHandler extends TriggerHandler {}");
+        Assert.assertSame(ASTUserClass.class, node.getClass());
+        ASTUserClass toplevel = (ASTUserClass) node;
+        Assert.assertEquals("TriggerHandler", toplevel.getSuperTypeName());
+    }
+
+    @Test
+    public void testSuperClassType2() {
+        ApexNode<?> node = parse("public class AccountTriggerHandler extends Other.TriggerHandler {}");
+        Assert.assertSame(ASTUserClass.class, node.getClass());
+        ASTUserClass toplevel = (ASTUserClass) node;
+        Assert.assertEquals("Other.TriggerHandler", toplevel.getSuperTypeName());
+    }
 }